+++ ygo-service/api/health_handler.go
@@ -12,5 +12,5 @@ func (s *healthServiceServer) APIStatus(ctx context.Context, req *emptypb.Empty)
 	logger, _ := util.NewRequestSetup(context.Background(), "Status")
 	logger.Info("Retrieving status of gRPC service")
 
-	return &health.APIStatusDetails{Version: "1.1.0"}, nil
+	return &health.APIStatusDetails{Version: "1.2.0"}, nil
 }
